# backend/src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Amana Backend API",
    "version": "1.0.0",
...
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Current versioned lane (recommended)"
    },
    {
      "url": "http://localhost:4000",
      "description": "Deprecated legacy alias (serves Deprecation/Sunset headers)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-secret",
        "description": "Secret key authentication for admin-only management routes."
      }
    },
...
    "schemas": {
...
      "StreamStatus": {
        "type": "string",
        "description": "Lifecycle state of a vested token stream. ACTIVE and SUSPENDED are live\nstates an admin may terminate from; TERMINATED and COMPLETED are terminal.",
        "enum": [
          "ACTIVE",
          "SUSPENDED",
          "TERMINATED",
          "COMPLETED"
        ]
      },
...
      "VestingState": {
        "type": "string",
        "description": "Derived from `claimed` vs. `totalVested` — independent of the stream's\nlifecycle `status`, since e.g. a SUSPENDED stream may already be\nfully vested.",
        "enum": [
          "not_started",
          "vesting",
          "fully_vested"
        ]
      },
...
      "AdminStreamListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminStreamSummary"
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              }
            },
            "required": [
              "page",
              "limit",
              "total",
              "totalPages"
            ]
          }
        },
        "required": [
          "items",
          "pagination"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/admin/streams": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List admin-accessible streams (#51)",
        "description": "Paginated list of vested token streams, newest first. Supports\nfiltering by lifecycle `status`, derived `vestingState` (how much of\nthe stream has actually vested, independent of lifecycle status),\nand `adminTag` (operator-assigned label). Backs the admin stream\nlist UI that the clawback confirmation flow (#56, #57) is launched\nfrom.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/StreamStatus"
            }
          },
          {
            "name": "vestingState",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/VestingState"
            }
          },
          {
            "name": "adminTag",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated stream summaries.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStreamListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/LegacyValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/admin/streams/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get stream state by ID",
        "description": "Retrieve cached stream state for admin queries. Returns cached result\nwhen fresh, otherwise fetches from DB and caches it.\n\nError code references: `NOT_FOUND` (unknown stream), `VALIDATION_ERROR` (invalid stream ID format).",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/StreamIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Stream state.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid stream ID format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/admin/streams/{id}/clawback/preview": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Preview a stream clawback before execution",
        "description": "Validates a requested clawback amount against the stream's actual\nremaining vested (unclaimed) balance and returns the resulting\npost-clawback balance, without performing the clawback. Powers the\nadmin confirmation modal (#56) and client-side amount validation\n(#57) — the frontend calls this before letting the operator confirm.\n\nConcurrent clawback requests against the same stream are serialized;\na request against a stream with a clawback already in flight gets\n`409`. All routes in this section return `404` when the\n`ADMIN_ROUTES_ENABLED` feature flag is off.\n\nError code references: `NOT_FOUND` (unknown stream),\n`CLAWBACK_INVALID_AMOUNT` (zero/negative amount),\n`CLAWBACK_TOO_LARGE` (amount exceeds `remainingVested`),\n`DOMAIN_ERROR` (clawback already in progress for this stream).",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/StreamIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamClawbackPreviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Clawback preview computed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamClawbackPreviewResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid stream ID format, invalid amount, or amount exceeds the remaining vested balance.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "409": {
            "description": "A clawback is already in progress for this stream.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppErrorResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/admin/streams/{id}/terminate": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Terminate a stream (admin only)",
        "description": "Permanently terminates a vested token stream. Admin only.\n\nThe stream's state is validated before anything is signed or written:\nan unknown stream is a `404`, and a stream already in a terminal state\n(`TERMINATED`, `COMPLETED`) is a `409` rather than a silent no-op, so a\ndouble-submit is visible to the operator instead of producing a second\naudit record for one real action.\n\nSupplying `unsignedTxXdr` signs the on-chain terminate call with the\nconfigured admin key and returns the signed XDR for submission; the\nresulting on-chain event is ingested by the event listener. Omitting it\nperforms the backend state transition only. Every successful\ntermination writes an `AdminActionAudit` record (`STREAM_TERMINATE`).\n\nTermination is irreversible — `reversible` is always `false`.\nError code references: `NOT_FOUND`, `DOMAIN_ERROR`, `VALIDATION_ERROR`.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/StreamIdPath"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamTerminateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stream terminated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamTerminateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid stream ID format or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "409": {
            "description": "Stream is not in a state that can be terminated, or the stream is locked for maintenance.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppErrorResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/admin/streams/{id}/lock": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Lock a stream for maintenance (admin only)",
        "description": "Marks a stream as locked for operational maintenance. Idempotent —\nlocking an already-locked stream returns 200 without changing state.\n\nLocked streams reject other admin mutations (suspend, resume,\nterminate, clawback) with `409 Conflict` until unlocked.\nUser-facing operations are unaffected.\n\nEvery successful lock writes an `AdminActionAudit` record\n(`STREAM_LOCK`). Error code references: `NOT_FOUND`, `VALIDATION_ERROR`.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/StreamIdPath"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamLockRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stream locked (or already locked).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamLockResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid stream ID format or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/admin/streams/{id}/unlock": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Unlock a previously locked stream (admin only)",
        "description": "Removes the maintenance lock from a stream. Idempotent — unlocking an\nalready-unlocked stream returns 200 without changing state.\n\nEvery successful unlock writes an `AdminActionAudit` record\n(`STREAM_UNLOCK`). Error code references: `NOT_FOUND`, `VALIDATION_ERROR`.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/StreamIdPath"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamUnlockRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stream unlocked (or already unlocked).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamUnlockResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid stream ID format or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/admin/streams/{id}/reconcile": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Reconcile a stream's backend state against on-chain events",
        "description": "Compares a stream's backend database record against ingested on-chain\ncontract events (StreamClawbackEvent, ProcessedEvent) and admin audit\nactions, returning a structured reconciliation result.\n\nFields compared:\n- `pendingClawback`: DB value vs sum of on-chain StreamClawbackEvent amounts\n- `unclaimed`: computed (totalVested - claimed - onChainClawbacks) vs DB value\n- `status`: checks if an admin terminate audit record exists but the stream is still live\n- `claimed`: flags when on-chain clawbacks exist but claimed is zero\n\nThe response includes a snapshot of the current backend state and an\non-chain summary so the operator can see both sides at a glance.\nAdmin only. No request body required.\nError code references: `NOT_FOUND`.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Stream identifier.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reconciliation completed — check `consistent` field for result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamReconciliationResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/api/admin/streams/{id}/suspend": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Suspend a vesting stream",
        "description": "Suspend a stream without immediate clawback. Gated by `ADMIN_ROUTES_ENABLED`.\nError code references: `AUTH_ERROR`, `VALIDATION_ERROR`, `INTERNAL_ERROR`.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/StreamIdPath"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "maxLength": 500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stream suspended."
          },
          "400": {
            "description": "Invalid stream ID format or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppErrorResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  }
}